+++ services/api/dataApi/requirements-test.txt
@@ -1,1 +1,2 @@
+pyarrow
 file:../../../packages/testing
