+++ services/api/dataApi/tests/test_reports.py
@@ -1,81 +1,36 @@
-from datetime import datetime
-from unittest.mock import patch
-
 import pytest
-from common.repositories.reports import ReportsRepository
+import app
+from moto import mock_sqs
+import json
 
 
-def test_get(dynamo_mock):
-    with ReportsRepository() as repo:
-        get_result = repo.get("testReport")
-    assert get_result['name'] == 'testReport'
-    assert get_result['dataProtection'] == 3
-    assert get_result['queryString'] == "COUNT(*) from dev_level_3_database.cv_partner_employees"
+@pytest.fixture
+def client():
+    app.app.config['TESTING'] = True
+
+    with app.app.test_client() as client:
+        yield client
 
 
-def test_get_by_one_table(dynamo_mock):
-    with ReportsRepository() as repo:
-        get_by_tables_result = repo.get_by_tables(["test_table1"])
-    assert len(get_by_tables_result) == 1
-    assert get_by_tables_result[0]["name"] == "testReport"
+def test_delete_report(client, dynamo_mock):
+    response = client.delete('/data/report/testReport')
+    items = dynamo_mock.scan()['Items']
+    assert response.status_code == 204
+    assert all(item['name'] != 'testReport' for item in items)
 
 
-def test_get_by_many_tables(dynamo_mock):
-    with ReportsRepository() as repo:
-        get_by_tables_result = repo.get_by_tables(["test_table1", "test_table2"])
-    assert len(get_by_tables_result) == 1
-    assert get_by_tables_result[0]["name"] == "testReport"
+def test_delete_report_publish_event(client, dynamo_mock, sns_topic, sqs_queue):
+    sns_topic.subscribe(
+        Protocol='sqs',
+        Endpoint=sqs_queue.attributes['QueueArn'])
+    response = client.delete('data/report/testReport')
+    message = json.loads(json.loads(next(iter(sqs_queue.receive_messages())).body)['Message'])
+
+    assert response.status_code == 204
+    assert message['name'] == 'testReport'
+    assert message['dataProtection'] == 3
 
 
-def test_all(dynamo_mock):
-    with ReportsRepository() as repo:
-        all_result = repo.all()
-    assert len(all_result) == 2
-    assert all_result[0]["name"] == "testReport"
-    assert all_result[1]["name"] == "anotherReport"
-
-
-def test_create(dynamo_mock):
-    additional_report = {
-        "dataProtection": 3,
-        "name": "new_report",
-        "queryString": "COUNT(*) from dev_level_3_database.cv_partner_employees",
-        "tables": [
-            "test_table5",
-            "test_table6"
-        ],
-        "lastUsed": "fail",
-        "lastCacheUpdate": "fail"
-    }
-    with ReportsRepository() as repo:
-        repo.create(additional_report)
-    new_item_result = dynamo_mock.get_item(Key={"name": "new_report"})
-    assert new_item_result['Item']['name'] == "new_report"
-    assert new_item_result['Item']['created'] is not None
-
-
-def test_update_cache_time(dynamo_mock):
-    with patch("common.repositories.reports.datetime") as mock_datetime:
-        mock_datetime.now.return_value = datetime(2020, 1, 1)
-        mock_datetime.side_effect = lambda *args, **kwargs: datetime(*args, **kwargs)
-        with ReportsRepository() as repo:
-            repo.update_cache_time("testReport")
-        new_item_result = dynamo_mock.get_item(Key={"name": "testReport"})
-        assert new_item_result['Item']['lastCacheUpdate'] == datetime(2020, 1, 1).isoformat()
-
-
-def test_update_cache_time_non_existing(dynamo_mock):
-    with patch("common.repositories.reports.datetime") as mock_datetime:
-        mock_datetime.now.return_value = datetime(2020, 1, 1)
-        mock_datetime.side_effect = lambda *args, **kwargs: datetime(*args, **kwargs)
-        with ReportsRepository() as repo:
-            with pytest.raises(KeyError):
-                repo.update_cache_time("testReport_non_existing")
-
-
-def test_delete(dynamo_mock):
-    with ReportsRepository() as repo:
-        repo.delete("testReport")
-    all_reports = dynamo_mock.scan()
-    assert len(all_reports['Items']) == 1
-    assert all_reports['Items'][0]['name'] == "anotherReport"
+def test_delete_report_not_found(client, dynamo_mock):
+    response = client.delete('data/report/fakeReport')
+    assert response.status_code == 404
